Add files via upload
This project includes two pages i.e. the first page which displays a scoreboard of  a competition and the second one about the details of the submissions. The pages was build using HTML ,CSS, Javascript and for the table i have used v-data-table. You can  sort the data as per your needs or filter it by name. The data for the scoreboard is currently hardcoded in javascript file, but in future update we can integrate a database or API to make it more dynamic.